Augment rb tree
An augmented rb tree is an rb tree that invokes a user-defined callback
on each node whose configuration changed as a result of a node insertion
or deletion (including rotated nodes). The callback is invoked in the
bottom-up manner, starting from the affected leaf node and going up to
the tree root. It can be used for maintaining per node data. The data
assigned to a node should be a function of the set of nodes that
constitute the sub-tree rooted at the node. Using an augmented tree as
the foundation, one can build more complex structures, such as interval
trees or weighted trees.

To generate an augmented rb_tree, use rb_gen_aug or rb_gen_ext_key_aug
macro. Apart from arguments taken by rb_gen and rb_gen_ext_key, these
macros take an augmentation callback, which is passed the node whose
value should be recomputed and its children. The macros do not affect
the generated function signatures.

In addition to rb_gen_aug and rb_gen_ext_key_aug, this patch also
introduces a new tree iterator. The iterator performs pre-order
depth-first tree traversal and allows to skip certain branches.
It is absolutely necessary to make use of tree augmentation.

To test tree augmentation, implement a simple weighted tree, which
maintains the size of the sub-tree rooted at each node. It allows to
efficiently compute the rank of each element. In the course of the
test build a random tree using both insertions and deletions and check
that tree properties hold and rank is computed correctly.