small: fix garbage collection
A typo in small_collect_garbage() resulted in totally broken garbage
collection of small objects. As a result of this bug users observe
constantly growing box.slab.info().items_used in Tarantool. Fix it
and add a unit test to make sure we will not step on this in future.

See https://github.com/tarantool/tarantool/issues/2842