small: fix false-positive 'out of array bound' error
Static allocator uses a global char[] array to
hand it out in chunks. The array addressing was done
via '&buffer[index]', but GCC compiler thinks that
there is an out of bound array access, when sees
something like static_alloc(max_size + 1).

Having static_alloc() inlined it somewhy thinks, that
this size is used as an index of the static buffer,
despite the fact, that such sizes are filtered out in
'if's before indexing the buffer.

This test was leading to the compilation error:

    test/static.c:78:
    static_alloc(SMALL_STATIC_SIZE + 1)

Static_alloc checks if a current position + size >
max_size, then check is size > max_size, and then
returns NULL, the test checks it. But GCC can't detect
that, and thinks that this code makes
'buffer[max_size + 1]'.

This commit uses pointer arithmetic instead of address
operator to avoid this bogus error.