slab_arena: Enhance slab_arena_create to support madvise hints
The purpose of this extension is to be able to eliminate
some of slab arenas from dumping, especially when huge
amount of memory involves.

To keep backward compatibility we map flags passed into
new SLAB_ARENA_x flags. To be able to distinguish old
and new interfaces we reserve high bit of the integer
value and remap flags internally to syscall compatible
form.

Same time I have to modify build script to define
_GNU_SOURCE symbol otherwise madvise bits won't be
found.

Another aspect of the patch is simple feature engine:
we will need to test if dontdump is supported runtime,
for this sake one can call

	small_test_feature(SMALL_FEATURE_DONTDUMP);

to test if we can exclude particular slab arena
from dumping.

Part-of https://github.com/tarantool/tarantool/issues/3509