build: export small/ibuf.c symbols
Tarantool is going to change the way how symbols are exported for
FFI and the public C API. That required from small/ explicitly
give a list of symbols, which should be exposed.

The patch does it for ibuf. All the other symbols are not needed
to be visible. Ibuf is intensively used by FFI.

Part of tarantool/2971