region: new region_alloc_array, updated alloc_object
Region's API provided a macro for aligned allocation of an object:
region_alloc_object(). But not less frequent there is a case when
need to allocate an array of objects, also aligned. The patch adds
region_alloc_array() for that.

Also the patch adds an out parameter 'size' for both macros. It
simplifies total size calculation, which is needed almost always,
because total size is included into an error message, if the
allocation fails.

Needed for https://github.com/tarantool/tarantool/issues/4609