test: fix build of mempool test
Since fcac155d mempool.c depends on trivia/util.h, that is not a
part of small library but a part of tarantool. Actually there's
only one thing that is needed from the header - alignof macro.
Extract it and make small library independent from tarantool.

This issue has been fixed but this patch is more correct