small: implement new size class evaluation
small_alloc uses a collection of mempools of different sizes.
If small_alloc stores all mempools in an array then it have to determine
an offset in that array where the most suitable mempool is.
Let's name the offset as 'size class' and the size that the corresponding
mempool allocates as 'class size'.
Historically the class sizes grow incrementally up to some point and then
(at some size class) grow exponentially with user-provided factor.
Apart from incremental part the exponential part is not very obvious.
Binary search and floating-point logarithm could be used for size class
determination but both approaches seem to be too slow.

This patch implements faster size class evaluation based on integral
operations.

Part of #5216