small: add granularity option to small_alloc_create
Granularity is an option that allows user to set
multiplicity of memory allocation in small allocator.
Granulatiry must be power of two and >= sizeof(intptr_t)
in case when we used small allocator for all allocation
types and >= 4 in case when we used it for packed structs,
or types which size <= 4.