small: performance test refactoring
Test is completely redesigned, now it used google benchmark,
instead of custom performance metrics. Test checks the performance
of memory allocation and deallocation operations for typical workload.
First of all test allocates 250000 objects with size ranging from 20
to 100 bytes, or 1000000 objects with size ranging from 1000 to 10000
bytes and push it in the vector.
Then, in a loop, test allocates memory for an object of the appropriate
size, push it in the vector, and frees up memory for a random object in
the vector. Test calculates and print time of one pair of memory allocation
and deallocation operations and also count of operations.
Test also checks performance for different alloc_factor and slab_size.