build: fix tarantool build failure on xcode 12.5
`VERSION` file in small is treanted as C++ standard library on a
filesystem with case-insensitive names. So we create `include`
subfolder and then `include/small` subfolder in this project and
move all headers from `small` subfolder to `include/small`. Then
we change `include_directories` in tarantool CMake from the root
of this project to `include` subfolder. Also we rename `features.h`
to `small_features.h` because of name conflict with file from standart
c library.
Part of #6076