Fix incorrect mempool group creation
If slab order of first several pools was equal to zero,
all these pools added to same group with pools with slab
order is equal to one. It's leds to incorrect behavior
of `smfree` function, because this function relies on
same slab order size of all pools in mempool group.
Fix this and add assertion and test to check it.