rlist: fix ASAN error for aligned struct
During a list traversal, we check if the current item points
to the list head with &item->member. If the type of the item or
the member is supposed to be aligned, it will result in an ASAN
error:

  runtime error: member access within misaligned address 0x000001a3f550
  for type 'struct iproto_connection', which requires 64 byte alignment

Fix this by comparing an item to a list head using rlist_entry on
the head instead.