matras: introduce statistics
This commit adds the new matras_stat struct which is supposed to be
passed to matras_create() for extent accounting. It's okay to pass the
same struct to different matras allocators, in which case the statistics
will be accumulated. Currently, matras_stat includes two counters: the
total number of extents and the number of extents used by read views.
We need this to implement read view statistics in Tarantool.

Needed for https://github.com/tarantool/tarantool-ee/issues/143