asan: add obuf ASAN implementation
If ENABLE_ASAN build flag is set then obuf implementation is changed.
This implementation is ASAN friendly that is it allows to do buffer
overflow/underflow and use-after-free checks as well as leak checks.

Also while we at it let's drop obuf_capacity. It is not used and a bit
strange.

Part of https://github.com/tarantool/tarantool/issues/7327