matras: fix matras_view::block_count overflow
Currently the maximum number of blocks in a single instance of matras is
limited by 2^32, however this number is 1 more than can be represented by
`matras_id_t block_count'. When a number of records in Tarantool's LIGHT
hash table reaches this limit, block_count overflows to 0, resulting in a
crash or data loss.
Even the assertion `matras_log2(m->head.block_count) < m->log2_capacity'
didn't help to detect a bug, because matras_log2(UINT32_MAX) is 31, while
log2_capacity is 32. This patch limits block_count to 2^31 and adds missed
`return NULL' when the maximum capacity is reached in matras_alloc() and
matras_alloc_range().

Needed for tarantool/tarantool#3594