debug: add assertions of ASAN version to non ASAN
ASAN allocators has more prerequisite assertions than regular
allocators. Thus if some of these assertions are violated then we will
catch it only lately on full-ci check. Besides that assertions in ASAN
are not easy to investigate because coredumps are not produced. Let's
add same prerequisite assertions to the debug build for the convenience.
We only add the assertions that are easy to check.

Follow-up https://github.com/tarantool/tarantool/issues/7327