matras: introduce the matras extent reservation
This commit intrduces a new entity: `matras_allocator`, which is a
wrapper over an external allocator implementing extent reservation
under the hood and allowing to share the reserved memory between
several matras instances. Now the `matras_create` consumes this
object as a parameter instead of `alloc_func`, `free_func`, `ctx`
and `extent_size`.

Also introduced the `matras_touch_reserve` function that can be used
for reserving extents based on the planned amount of blocks to call
`matras_touch` for.

Needed for tarantool/tarantool-ee#822