small: improve real_size calculation in small_alloc_info
Currently `real_size` is calculated as the size of a memory block that
would be allocated for the requested number of bytes. However, this
calculation is non-idempotent. For details see commit df7ba1f02f88
("small: implement new allocator strategy").

This patch switches `small_alloc_info()` to `slab_from_ptr()` that
returns the actual slab that contains the requested `ptr`, so now
`real_size` is always correct.

Needed for tarantool/tarantool#10217