lsregion: fix mixing cached and large slabs in lsregion_to_iovec
Cached slabs used to get their slab_id too early, at the moment of
caching, and not at the moment of using the cache.

This didn't affect anything for normal-sized slabs, because for them the
cache is always checked before trying to allocate another slab.

But in case a large slab was used, the invariant that slabs always
receive a monotonically growing slab_id was broken. First, a normal slab
was cached and received a slab_id = i. Then a large slab got allocated
and received a slab_id = i + 1. Later a cached slab was taken with
slab_id = i.

As a result mixing large and normal allocations could lead to a
situation when lsregion_to_iovec couldn't flush a normal slab, because
it considered it as already seen.

Let's fix this and always assign slab_ids at the moment the slabs are
first used (either taken from cache or allocated).

Closes #109